Refactor models into another file Add some parsing of BASC
