+++ requirements.txt
@@ -1,4 +1,3 @@
-pandas
 asn1crypto==0.23.0
 attrs==17.2.0
 Automat==0.6.0
@@ -38,4 +37,4 @@ w3lib==1.18.0
 webencodings==0.5.1
 Werkzeug==0.12.2
 zope.interface==4.4.3
-xlrd==1.1.0
+xlrd==1.1.0e
